refactor(core): Split out subworkflow policy check from permission checker (no-changelog)
